+++ crates/wasm/CHANGELOG.md
@@ -1,6 +1,17 @@
 # `moc-wasm` Change Log
 
-## 0.1.0-alpha
+## 0.2.0
+
+Released 2021-xx-xx.
+
+### Added
+
+* `from_valued_cells`: MOC from multi-order map
+* `split`: split a disjoint S-MOC into joint S-MOCs
+* `split_count`: count the number of joint S-MOCs in a disjoint S-MOC
+
+
+## 0.1.0
 
 Released 2021-10-18.